Like and Dislike Buttons - WITH TOAST :D
I tried to style as much as I could but AS doesn't give much to go on. They display a toast when clicked though!

Also some scaling fixes so that there wasn't a huge gap between the smaller paragraphs and the buttons. In other words, it now auto resizes to fit to scale.